# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: "¿Dónde me atienden? - API del agente vocal"
  version: 0.1.0
  description: |
    Contrato HTTP del servicio agente-vocal (Reto 01, Kognia Labs). El servicio
    es público y sin autenticación (ver docs/adr/0004-acceso-publico-sin-login.md).

    La conversación por voz usa un WebSocket. OpenAPI no modela WebSockets, por
    eso aquí solo se documenta el upgrade HTTP de /ws/agent y sus respuestas de
    rechazo; los mensajes, códigos de cierre y límites están en
    docs/api/websocket-protocol.md.

    Todas las respuestas llevan las cabeceras de seguridad definidas en
    web/next.config.mjs: Strict-Transport-Security, Content-Security-Policy-Report-Only,
    X-Content-Type-Options, X-Frame-Options, Referrer-Policy y Permissions-Policy
    (RNF-005). Las respuestas de rechazo del upgrade las escribe server.mjs
    directamente sobre el socket y no llevan esas cabeceras.
servers:
  - url: https://agente-vocal-583590264456.us-east1.run.app
    description: Producción (Cloud Run, us-east1)
  - url: http://localhost:3000
    description: Desarrollo local (npm run dev)
security: []
tags:
  - name: operacion
    description: Salud del servicio.
  - name: documento
    description: Carga del documento de la persona y su brief.
  - name: voz
    description: Conversación por voz.
paths:
  /api/health:
    get:
      tags: [operacion]
      operationId: getHealth
      summary: Verificación de salud
      description: Responde 200 si el proceso está sirviendo peticiones. No consulta proveedores externos.
      x-requirement: [RNF-003, RNF-001]
      responses:
        "200":
          description: El servicio está en línea.
          headers:
            Cache-Control:
              description: Siempre no-store.
              schema:
                type: string
                const: no-store
            Strict-Transport-Security:
              $ref: "#/components/headers/Strict-Transport-Security"
            X-Content-Type-Options:
              $ref: "#/components/headers/X-Content-Type-Options"
            X-Frame-Options:
              $ref: "#/components/headers/X-Frame-Options"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
              example:
                status: ok
  /api/document:
    post:
      tags: [documento]
      operationId: uploadDocument
      summary: Sube un documento y devuelve su brief
      description: |
        El cuerpo es el archivo en bruto (no multipart). El servidor detecta el
        tipo real por su firma: PDF (`%PDF-`), DOCX (ZIP con `word/document.xml`)
        o TXT (UTF-8 válido sin bytes nulos). La extensión y el Content-Type que
        envía el cliente no se usan para decidir el tipo. El nombre original no
        se recibe ni se guarda; el servidor genera `documentId` (UUID v4).

        El texto extraído se guarda solo en memoria de la instancia, como máximo
        20 000 caracteres (si hay más, `truncado` es true), durante 30 minutos.
        No hay base de datos ni disco. Con ese `documentId` se abre la sesión de
        voz en `/ws/agent?doc=<documentId>`.

        El brief lo genera DeepSeek con un plazo de 25 s. Si falla, la carga
        igual se acepta y `brief` es null: la persona puede hablar del documento.

        Controles: Origin en ALLOWED_ORIGINS (403), 5 cargas por minuto y 1
        simultánea por IP (429), 20 MB como máximo (413).
      x-requirement: [RF-001, RF-003, RNF-004]
      parameters:
        - name: Origin
          in: header
          required: true
          schema:
            type: string
            format: uri
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              maxLength: 20971520
      responses:
        "201":
          description: Documento aceptado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentCreated"
              example:
                documentId: 0b9f6c1e-3a52-4d0e-9a51-6f1d2b7c8e90
                tipo: pdf
                caracteres: 18342
                truncado: false
                brief:
                  resumen: Contrato de prestación de servicios de salud entre una EPS y una IPS de Leticia.
                  preguntas:
                    - ¿Qué servicios cubre el contrato?
                    - ¿Cuánto dura el contrato?
                    - ¿Qué pasa si la IPS incumple?
        "400":
          description: Cuerpo vacío (vacio).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: Origin ausente o fuera de la lista blanca (origen_no_permitido).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: Más de 20 MB (demasiado_grande).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: La firma no es PDF, DOCX ni TXT, o el archivo está corrupto (tipo_no_soportado).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: El archivo es válido pero no tiene texto extraíble, por ejemplo un PDF escaneado (sin_texto).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Más de 5 cargas por minuto o una carga en curso desde la IP (demasiadas_cargas).
          headers:
            Retry-After:
              schema:
                type: integer
                const: 60
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Fallo inesperado del servidor (error_interno). El detalle queda solo en los logs.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /ws/agent:
    get:
      tags: [voz]
      operationId: openVoiceSession
      summary: Abre una sesión de voz (upgrade a WebSocket)
      description: |
        Upgrade HTTP a WebSocket. Tras el 101, el protocolo de mensajes está en
        docs/api/websocket-protocol.md.

        Si viene `doc`, la sesión usa ese documento como contexto del agente;
        un `doc` con formato inválido, desconocido o vencido se rechaza con 404
        antes de consumir cupo.

        Controles de admisión, en este orden: Origin en la lista blanca
        ALLOWED_ORIGINS (403), documento (404), tasa de 10 intentos por minuto por IP (429),
        cupo global de 8 sesiones por instancia (503) y cupo de 2 sesiones
        simultáneas por IP (429). La IP del cliente es la última entrada de
        X-Forwarded-For. Los rechazos tienen cuerpo vacío y cierran la conexión.
      x-requirement: [RF-002, RF-004, RF-005, RF-009, RF-019, RNF-004]
      x-protocol: websocket
      x-protocol-doc: docs/api/websocket-protocol.md
      parameters:
        - name: doc
          in: query
          required: false
          description: documentId devuelto por POST /api/document.
          schema:
            type: string
            format: uuid
        - name: Origin
          in: header
          required: true
          description: Debe coincidir exactamente con un origen de ALLOWED_ORIGINS.
          schema:
            type: string
            format: uri
          example: https://agente-vocal-583590264456.us-east1.run.app
        - name: Upgrade
          in: header
          required: true
          schema:
            type: string
            const: websocket
        - name: Sec-WebSocket-Version
          in: header
          required: true
          schema:
            type: string
            const: "13"
        - name: Sec-WebSocket-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        "101":
          description: Upgrade aceptado. Comienza la sesión WebSocket.
        "403":
          description: Origin ausente o fuera de la lista blanca (motivo bad_origin).
        "404":
          description: El parámetro doc no corresponde a un documento vigente (doc_not_found).
        "429":
          description: Más de 10 intentos por minuto desde la IP (rate_limited) o ya hay 2 sesiones activas de la IP (ip_cap).
          headers:
            Retry-After:
              description: Segundos a esperar antes de reintentar.
              schema:
                type: integer
                const: 60
        "503":
          description: La instancia alcanzó el cupo global de sesiones (global_cap).
components:
  schemas:
    Health:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          const: ok
    DocumentCreated:
      type: object
      additionalProperties: false
      required: [documentId, tipo, caracteres, truncado, brief]
      properties:
        documentId:
          type: string
          format: uuid
        tipo:
          type: string
          enum: [pdf, docx, txt]
        caracteres:
          type: integer
          minimum: 1
          maximum: 20000
          description: Caracteres de texto guardados para la conversación.
        truncado:
          type: boolean
          description: true si el documento tenía más de 20 000 caracteres y se cortó.
        brief:
          oneOf:
            - $ref: "#/components/schemas/Brief"
            - type: "null"
    Brief:
      type: object
      additionalProperties: false
      required: [resumen, preguntas]
      properties:
        resumen:
          type: string
          minLength: 1
          maxLength: 600
        preguntas:
          type: array
          minItems: 3
          maxItems: 5
          items:
            type: string
            minLength: 1
            maxLength: 200
    Error:
      type: object
      additionalProperties: false
      required: [error, mensaje]
      properties:
        error:
          type: string
          enum: [vacio, origen_no_permitido, demasiado_grande, tipo_no_soportado, sin_texto, demasiadas_cargas, error_interno]
        mensaje:
          type: string
          description: Texto en español para mostrar a la persona. Nunca incluye detalles internos.
  headers:
    Strict-Transport-Security:
      schema:
        type: string
        const: max-age=63072000; includeSubDomains
    X-Content-Type-Options:
      schema:
        type: string
        const: nosniff
    X-Frame-Options:
      schema:
        type: string
        const: DENY